Finish the simple version of the site
Added a quick description about the convex hull and what it is. Then on the convex hull
page make it so that there is a convex hull on initial page load. And then when the user
hits new graph we delete the older points and find the convex hull of a new data set.